Limit violations refactoring
