+++ src/main/java/com/powsybl/openloadflow/sa/AbstractSecurityAnalysis.java
@@ -9,33 +9,21 @@ package com.powsybl.openloadflow.sa;
 import com.powsybl.computation.CompletableFutureTask;
 import com.powsybl.computation.ComputationManager;
 import com.powsybl.contingency.ContingenciesProvider;
-import com.powsybl.iidm.network.Branch;
-import com.powsybl.iidm.network.LimitType;
 import com.powsybl.iidm.network.Network;
-import com.powsybl.loadflow.LoadFlowParameters;
 import com.powsybl.math.matrix.MatrixFactory;
-import com.powsybl.openloadflow.OpenLoadFlowParameters;
 import com.powsybl.openloadflow.graph.GraphDecrementalConnectivityFactory;
 import com.powsybl.openloadflow.network.LfBranch;
 import com.powsybl.openloadflow.network.LfBus;
-import com.powsybl.openloadflow.network.LfNetwork;
-import com.powsybl.openloadflow.network.util.ActivePowerDistribution;
-import com.powsybl.openloadflow.util.PerUnit;
-import com.powsybl.security.LimitViolation;
-import com.powsybl.security.LimitViolationType;
 import com.powsybl.security.SecurityAnalysisParameters;
 import com.powsybl.security.SecurityAnalysisReport;
 import com.powsybl.security.monitor.StateMonitor;
 import com.powsybl.security.monitor.StateMonitorIndex;
-import org.apache.commons.lang3.tuple.Pair;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.util.List;
-import java.util.Map;
 import java.util.Objects;
 import java.util.concurrent.CompletableFuture;
-import java.util.stream.Stream;
 
 /**
  * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
@@ -78,157 +66,4 @@ public abstract class AbstractSecurityAnalysis {
 
     abstract SecurityAnalysisReport runSync(String workingVariantId, SecurityAnalysisParameters securityAnalysisParameters, ContingenciesProvider contingenciesProvider,
                                             ComputationManager computationManager);
-
-    /**
-     * Detect violations on branches and on buses
-     * @param branches branches on which the violation limits are checked
-     * @param buses buses on which the violation limits are checked
-     * @param violations list on which the violation limits encountered are added
-     */
-    protected void detectViolations(Stream<LfBranch> branches, Stream<LfBus> buses, Map<Pair<String, Branch.Side>, LimitViolation> violations) {
-        // Detect violation limits on branches
-        branches.forEach(branch -> detectBranchViolations(branch, violations));
-
-        // Detect violation limits on buses
-        buses.forEach(bus -> detectBusViolations(bus, violations));
-    }
-
-    /**
-     * Detect violation limits on one branch and add them to the given list
-     * @param branch branch of interest
-     * @param violations list on which the violation limits encountered are added
-     */
-    protected void detectBranchViolations(LfBranch branch, Map<Pair<String, Branch.Side>, LimitViolation> violations) {
-        // detect violation limits on a branch
-        // Only detect the most serious one (findFirst) : limit violations are ordered by severity
-        if (branch.getBus1() != null) {
-            branch.getLimits1(LimitType.CURRENT).stream()
-                .filter(temporaryLimit1 -> branch.getI1().eval() > temporaryLimit1.getValue())
-                .findFirst()
-                .map(temporaryLimit1 -> createLimitViolation1(branch, temporaryLimit1, LimitViolationType.CURRENT, PerUnit.ib(branch.getBus1().getNominalV()), branch.getI1().eval()))
-                .ifPresent(limitViolation -> violations.put(getSubjectSideId(limitViolation), limitViolation));
-
-            branch.getLimits1(LimitType.ACTIVE_POWER).stream()
-                  .filter(temporaryLimit1 -> branch.getP1().eval() > temporaryLimit1.getValue())
-                  .findFirst()
-                  .map(temporaryLimit1 -> createLimitViolation1(branch, temporaryLimit1, LimitViolationType.ACTIVE_POWER, PerUnit.SB, branch.getP1().eval()))
-                  .ifPresent(limitViolation -> violations.put(getSubjectSideId(limitViolation), limitViolation));
-
-            //Apparent power is not relevant for fictitious branches and may be NaN
-            double apparentPower1 = branch.computeApparentPower1();
-            if (!Double.isNaN(apparentPower1)) {
-                branch.getLimits1(LimitType.APPARENT_POWER).stream()
-                      .filter(temporaryLimit1 -> apparentPower1 > temporaryLimit1.getValue())
-                      .findFirst()
-                      .map(temporaryLimit1 -> createLimitViolation1(branch, temporaryLimit1, LimitViolationType.APPARENT_POWER, PerUnit.SB, apparentPower1))
-                      .ifPresent(limitViolation -> violations.put(getSubjectSideId(limitViolation), limitViolation));
-            }
-
-        }
-        if (branch.getBus2() != null) {
-            branch.getLimits2(LimitType.CURRENT).stream()
-                .filter(temporaryLimit2 -> branch.getI2().eval() > temporaryLimit2.getValue())
-                .findFirst() // only the most serious violation is added (the limits are sorted in descending gravity)
-                .map(temporaryLimit2 -> createLimitViolation2(branch, temporaryLimit2, LimitViolationType.CURRENT, PerUnit.ib(branch.getBus2().getNominalV()), branch.getI2().eval()))
-                .ifPresent(limitViolation -> violations.put(getSubjectSideId(limitViolation), limitViolation));
-
-            branch.getLimits2(LimitType.ACTIVE_POWER).stream()
-                  .filter(temporaryLimit2 -> branch.getP2().eval() > temporaryLimit2.getValue())
-                  .findFirst()
-                  .map(temporaryLimit2 -> createLimitViolation2(branch, temporaryLimit2, LimitViolationType.ACTIVE_POWER, PerUnit.SB, branch.getP2().eval()))
-                  .ifPresent(limitViolation -> violations.put(getSubjectSideId(limitViolation), limitViolation));
-
-            //Apparent power is not relevant for fictitious branches and may be NaN
-            double apparentPower2 = branch.computeApparentPower2();
-            if (!Double.isNaN(apparentPower2)) {
-                branch.getLimits2(LimitType.APPARENT_POWER).stream()
-                      .filter(temporaryLimit2 -> apparentPower2 > temporaryLimit2.getValue())
-                      .findFirst()
-                      .map(temporaryLimit2 -> createLimitViolation2(branch, temporaryLimit2, LimitViolationType.APPARENT_POWER, PerUnit.SB, apparentPower2))
-                      .ifPresent(limitViolation -> violations.put(getSubjectSideId(limitViolation), limitViolation));
-            }
-        }
-    }
-
-    protected static LimitViolation createLimitViolation1(LfBranch branch, LfBranch.LfLimit temporaryLimit1,
-                                                          LimitViolationType type, double scale, double value) {
-        return new LimitViolation(branch.getId(), type, null,
-                temporaryLimit1.getAcceptableDuration(), temporaryLimit1.getValue() * scale,
-                (float) 1., value * scale, Branch.Side.ONE);
-    }
-
-    protected static LimitViolation createLimitViolation2(LfBranch branch, LfBranch.LfLimit temporaryLimit2,
-                                                          LimitViolationType type, double scale, double value) {
-        return new LimitViolation(branch.getId(), type, null,
-                temporaryLimit2.getAcceptableDuration(), temporaryLimit2.getValue() * scale,
-                (float) 1., value * scale, Branch.Side.TWO);
-    }
-
-    protected static Pair<String, Branch.Side> getSubjectSideId(LimitViolation limitViolation) {
-        return Pair.of(limitViolation.getSubjectId(), limitViolation.getSide());
-    }
-
-    /**
-     * Detect violation limits on one branch and add them to the given list
-     * @param bus branch of interest
-     * @param violations list on which the violation limits encountered are added
-     */
-    protected void detectBusViolations(LfBus bus, Map<Pair<String, Branch.Side>, LimitViolation> violations) {
-        // detect violation limits on a bus
-        double scale = bus.getNominalV();
-        double busV = bus.getV();
-        if (!Double.isNaN(bus.getHighVoltageLimit()) && busV > bus.getHighVoltageLimit()) {
-            LimitViolation limitViolation1 = new LimitViolation(bus.getVoltageLevelId(), LimitViolationType.HIGH_VOLTAGE, bus.getHighVoltageLimit() * scale,
-                    (float) 1., busV * scale);
-            violations.put(getSubjectSideId(limitViolation1), limitViolation1);
-        }
-        if (!Double.isNaN(bus.getLowVoltageLimit()) && busV < bus.getLowVoltageLimit()) {
-            LimitViolation limitViolation2 = new LimitViolation(bus.getVoltageLevelId(), LimitViolationType.LOW_VOLTAGE, bus.getLowVoltageLimit() * scale,
-                    (float) 1., busV * scale);
-            violations.put(getSubjectSideId(limitViolation2), limitViolation2);
-        }
-    }
-
-    public static void distributedMismatch(LfNetwork network, double mismatch, LoadFlowParameters loadFlowParameters,
-                                           OpenLoadFlowParameters openLoadFlowParameters) {
-        if (loadFlowParameters.isDistributedSlack() && Math.abs(mismatch) > 0) {
-            ActivePowerDistribution activePowerDistribution = ActivePowerDistribution.create(loadFlowParameters.getBalanceType(), openLoadFlowParameters.isLoadPowerFactorConstant());
-            activePowerDistribution.run(network, mismatch);
-        }
-    }
-
-    /**
-     * Compares two limit violations
-     * @param violation1 first limit violation
-     * @param violation2 second limit violation
-     * @return true if violation2 is weaker than or equivalent to violation1, otherwise false
-     */
-    protected static boolean violationWeakenedOrEquivalent(LimitViolation violation1, LimitViolation violation2,
-                                                           SecurityAnalysisParameters.IncreasedViolationsParameters violationsParameters) {
-        if (violation2 != null && violation1.getLimitType() == violation2.getLimitType()) {
-            if (violation2.getLimit() < violation1.getLimit()) {
-                // the limit violated is smaller hence the violation is weaker, for flow violations only.
-                // for voltage limits, we have only one limit by limit type.
-                return true;
-            }
-            if (violation2.getLimit() == violation1.getLimit()) {
-                // the limit violated is the same: we consider the violations equivalent if the new value is close to previous one.
-                if (isFlowViolation(violation2)) {
-                    return Math.abs(violation2.getValue()) <= Math.abs(violation1.getValue()) * (1 + violationsParameters.getFlowProportionalThreshold());
-                } else if (violation2.getLimitType() == LimitViolationType.HIGH_VOLTAGE) {
-                    double value = Math.min(violationsParameters.getHighVoltageAbsoluteThreshold(), violation1.getValue() * violationsParameters.getHighVoltageProportionalThreshold());
-                    return violation2.getValue() <= violation1.getValue() + value;
-                } else if (violation2.getLimitType() == LimitViolationType.LOW_VOLTAGE) {
-                    return violation2.getValue() >= violation1.getValue() - Math.min(violationsParameters.getLowVoltageAbsoluteThreshold(), violation1.getValue() * violationsParameters.getLowVoltageProportionalThreshold());
-                } else {
-                    return false;
-                }
-            }
-        }
-        return false;
-    }
-
-    protected static boolean isFlowViolation(LimitViolation limit) {
-        return limit.getLimitType() == LimitViolationType.CURRENT || limit.getLimitType() == LimitViolationType.ACTIVE_POWER || limit.getLimitType() == LimitViolationType.APPARENT_POWER;
-    }
 }
